fully finctional code with stub in token area. honestly I don't understand how to pass dynamic data to this shit

+++ allure-exercise/src/main/java/qa/guru/allure/Issue.java
@@ -2,20 +2,15 @@ package qa.guru.allure;
 
 public class Issue {
     private long id;
-
-
     private int number;
     private String title;
     private String body;
 
-    public long getNumber() {
-        return number;
-    }
+    public int getNumber() {  return number;    }
 
     public void setNumber(int number) {
         this.number = number;
     }
-
 
     public long getId() {
         return id;
